pyproject.toml
[tool.poetry]
name = "nonebot_plugin_reboot"
version = "0.1.4"
description = "Reboot your bot by using command"
license = "MIT"
authors = ["18870 <a20110123@163.com>"]
readme = "README.md"
repository = "https://github.com/18870/nonebot-plugin-reboot"
keywords = ["nonebot"]

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0.0"
nonebot2 = ">=2.1.0, <3.0.0"
nonebot-adapter-onebot = { version = "^2.1.0", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
